Finished organizing methods to their respective modules
