+++ lib/stat_tracker.rb
@@ -9,9 +9,11 @@ require_relative './league_statistics'
 require_relative './league_helpers'
 require_relative './team_statistics'
 require_relative './team_helpers'
+require_relative './team_secondary_helpers'
 require_relative './team_seasonal_summary_helpers'
 require_relative './season_statistics'
 require_relative './season_helpers'
+
 
 require 'pry'
 
@@ -22,6 +24,7 @@ class StatTracker
   include TeamStatistics
   include TeamStatisticsHelpers
   include TeamSeasonalSummaryHelpers
+  include TeamSecondaryHelpers
   include SeasonStatistics
   include SeasonStatisticsHelpers
